rms/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
django-allauth==0.51.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.3.0
djangorestframework==3.14.0
filelock==3.8.2
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
platformdirs==2.6.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.6
Unidecode==0.4.21
urllib3==1.26.13
